move html from my-messages to the corresponding cards
