+++ src/app/my-messages/abstract-community-notification.ts
@@ -2,5 +2,5 @@ import { AbstractNotification } from './abstract-notification';
 
 export abstract class AbstractCommunityNotification extends AbstractNotification {
 
-  abstract isToDoType(): boolean;
+  abstract getCommunityInfo(): string;
 }
