+++ src/app/my-messages/common-message-card/common-message-card.component.html
@@ -1,3 +1,17 @@
-<p>
-  common-message-caard works!
-</p>
+<div fxFlex>
+  <div fxLayout="row">
+    <div class="messages-description-label">Type:</div>
+    <div>{{notification.getTypeText()}}</div>
+  </div>
+  <div fxLayout="row">
+    <div class="messages-description-label">Date:</div>
+    <div>{{notification.creationDatetime | date:'dd.MM.yyyy HH:mm:ss'}}</div>
+  </div>
+  <div fxLayout="row">
+    <div class="messages-description-label">Message:</div>
+    <div [innerHTML]="notification.getHtmlMessageText()"></div>
+  </div>
+</div>
+<div fxFlex="none">
+  <ng-content select="[buttonContent]"></ng-content>
+</div>
