+++ src/app/my-messages/common-message-card/common-message-card.component.scss
@@ -1,0 +1,5 @@
+@import "variables";
+
+.messages-description-label {
+  @extend %base-description-label;
+}
