+++ src/app/my-messages/common-message-card/common-message-card.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
+import { AbstractNotification } from '../abstract-notification';
 
 @Component({
   selector: 'app-common-message-card',
@@ -6,6 +7,8 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./common-message-card.component.scss']
 })
 export class CommonMessageCardComponent implements OnInit {
+
+  @Input() notification: AbstractNotification;
 
   constructor() { }
