+++ src/app/my-messages/community-message-card/community-message-card.component.scss
@@ -1,0 +1,5 @@
+@import "variables";
+
+.messages-description-label {
+  @extend %base-description-label;
+}
