+++ src/app/my-messages/member-kicked-out-notification.ts
@@ -1,4 +1,3 @@
-import { AbstractNotification } from './abstract-notification';
 import { Community } from '../communities/community';
 import { AbstractCommunityNotification } from './abstract-community-notification';
 
@@ -10,11 +9,12 @@ export class MemberKickedOutNotification extends AbstractCommunityNotification {
     return 'The user was kicked out of the community';
   }
 
-  isToDoType(): boolean {
-    return false;
+  getHtmlMessageText(): string {
+    return 'The user was kicked out of the community';
   }
 
-  getHtmlMessageText(): string {
-    return '';
+  getCommunityInfo(): string {
+    return this.community ?`<a href="/communities/${this.community.id}">${this.community.title}</a>`
+      : this.communityName;
   }
 }
