+++ src/app/my-messages/member-left-community-notification.ts
@@ -1,4 +1,3 @@
-import { AbstractNotification } from './abstract-notification';
 import { Community } from '../communities/community';
 import { User } from '../users/user';
 import { AbstractCommunityNotification } from './abstract-community-notification';
@@ -12,11 +11,16 @@ export class MemberLeftCommunityNotification extends AbstractCommunityNotificati
     return 'A member leaving a community';
   }
 
-  isToDoType(): boolean {
-    return false;
+  getHtmlMessageText(): string {
+    return 'A member leaving a community';
   }
 
-  getHtmlMessageText(): string {
-    return '';
+  getCommunityInfo(): string {
+    return this.community ?`<a href="/communities/${this.community.id}">${this.community.title}</a>`
+      : this.communityName;
+  }
+
+  getUser(): string {
+    return `${this.user.firstName} ${this.user.lastName}`;
   }
 }
