+++ src/app/my-messages/my-messages.component.scss
@@ -8,7 +8,4 @@
   @extend %base-error;
 }
 
-.messages-description-label {
-  @extend %base-description-label;
-}
