+++ src/app/my-messages/my-messages.component.ts
@@ -1,21 +1,14 @@
 import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { HttpErrorResponse } from '@angular/common/http';
 import { of } from 'rxjs';
-import { CommunityRegistrationService } from '../shared/community-registration.service';
-import { CommunityUserRegistration } from '../shared/community-user-registration';
 import { MessagesService } from './messages.service';
 import { DeleteConfirmationDialogComponent } from '../shared/delete-confirmation-dialog/delete-confirmation-dialog.component';
 import { MatDialog } from '@angular/material';
 import { AbstractNotification } from './abstract-notification';
-import { CommunityInvitationNotification } from './community-invitation-notification';
-import { JoinCommunityRequestNotification } from './join-community-request-notification';
 import { GlobalErrorHandlerService } from '../error/global-error-handler.service';
-import { Util } from '../util/util';
 import { UserIdentityService } from '../shared/user-identity.service';
 import { switchMap } from 'rxjs/operators';
 import { NotificationCounterService } from '../shared/notification-counter.service';
-import { InfoDialogComponent } from '../shared/info-dialog/info-dialog.component';
-import { TemplateLoaderService } from '../shared/template-loader.service';
 import { AbstractCommunityNotification } from './abstract-community-notification';
 
 @Component({
@@ -29,11 +22,9 @@ export class MyMessagesComponent implements OnInit {
   currentUserId: string;
   notifications$ = of([]);
 
-  constructor(private communityRegistrationService: CommunityRegistrationService,
-              private messageService: MessagesService,
+  constructor(private messageService: MessagesService,
               private userIdentityService: UserIdentityService,
               private notificationCounterService: NotificationCounterService,
-              private templateLoader: TemplateLoaderService,
               public dialog: MatDialog,
               private changeDetector: ChangeDetectorRef,
               private globalErrorHandlerService: GlobalErrorHandlerService) {
@@ -43,40 +34,8 @@ export class MyMessagesComponent implements OnInit {
     this.loadMessages();
   }
 
-  hasJoinRequestType<T extends AbstractNotification>(notification: T): boolean {
-    return notification.hasCommunityInvitationType() || notification.hasJoinCommunityType();
-  }
-
-  showAcceptDeclineButtons<T extends AbstractNotification>(notification: T) {
-    if (notification.hasCommunityInvitationType()) {
-      const invitationNotification: CommunityInvitationNotification = Util.createNotificationInstance(notification);
-      return invitationNotification.registration.user.id === this.currentUserId &&
-        (invitationNotification.registration.approvedByUser === null || invitationNotification.registration.approvedByCommunity === null);
-    } else if (notification.hasJoinCommunityType()) {
-      const joinRequestNotification: JoinCommunityRequestNotification = Util.createNotificationInstance(notification);
-      return joinRequestNotification.registration.user.id !== this.currentUserId &&
-        (joinRequestNotification.registration.approvedByUser === null || joinRequestNotification.registration.approvedByCommunity === null);
-    } else {
-      return false;
-    }
-  }
-
-  onAccept<T extends AbstractNotification>(notification: T) {
-    this.buildAcceptanceRequest(notification, true);
-  }
-
-  onDecline<T extends AbstractNotification>(notification: T) {
-    const dialogRef = this.dialog.open(DeleteConfirmationDialogComponent, {
-      width: '350px',
-      data: {
-        message: 'Are you sure you want to decline the request?'
-      }
-    });
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.buildAcceptanceRequest(notification, false);
-      }
-    });
+  isCommunityNotification<T extends AbstractNotification>(notification: T): boolean {
+    return (notification instanceof AbstractCommunityNotification);
   }
 
   showDeleteButton<T extends AbstractNotification>(notification: T): boolean {
@@ -107,53 +66,7 @@ export class MyMessagesComponent implements OnInit {
     });
   }
 
-  openWelcomeNotificationDialog() {
-    this.templateLoader.loadTemplate('/assets/templates/welcome-notification.html')
-      .subscribe(html => {
-        this.dialog.open(InfoDialogComponent, {
-          width: '550px',
-          data: {
-            message: html
-          }
-        });
-      }, errorResponse => {
-        this.handleErrorResponse(errorResponse);
-      });
-  }
-
-  private buildAcceptanceRequest<T extends AbstractNotification>(notification: T, isAccepted: boolean) {
-    if (notification.hasCommunityInvitationType()) {
-      const invitationNotification: CommunityInvitationNotification = Util.createNotificationInstance(notification);
-      const requestData = {
-        id: invitationNotification.registration.id,
-        approvedByUser: isAccepted,
-        approvedByCommunity: null
-      };
-
-      this.updateRegistration(invitationNotification.registration.community.id, requestData);
-    } else if (notification.hasJoinCommunityType()) {
-      const joinRequestNotification: JoinCommunityRequestNotification = Util.createNotificationInstance(notification);
-      const requestData = {
-        id: joinRequestNotification.registration.id,
-        approvedByUser: null,
-        approvedByCommunity: isAccepted
-      };
-
-      this.updateRegistration(joinRequestNotification.registration.community.id, requestData);
-    }
-  }
-
-  private updateRegistration(communityId: string, registration: CommunityUserRegistration) {
-    this.communityRegistrationService.updateRegistration(communityId, registration)
-      .subscribe(() => {
-        this.notifications$ = this.messageService.getUserNotifications(this.currentUserId);
-        this.notificationCounterService.decrementCount();
-      }, errorResponse => {
-        this.handleErrorResponse(errorResponse);
-      });
-  }
-
-  private loadMessages() {
+  public loadMessages() {
     this.notifications$ = this.userIdentityService.getUserIdentity()
       .pipe(switchMap(userIdentity => {
         this.currentUserId = userIdentity.userId;
@@ -161,14 +74,14 @@ export class MyMessagesComponent implements OnInit {
       }));
   }
 
-  private handleErrorResponse(errorResponse: HttpErrorResponse) {
+  public handleErrorResponse(errorResponse: HttpErrorResponse) {
     this.errorMessage = this.globalErrorHandlerService.createFullMessage(errorResponse);
     // Dirty fix because of: https://github.com/angular/angular/issues/17772
     this.changeDetector.markForCheck();
   }
 
   private isToDoType<T extends AbstractNotification>(notification: T): boolean {
-    return (notification instanceof AbstractCommunityNotification) && (notification.hasCommunityInvitationType() || notification.hasJoinCommunityType());
+    return this.isCommunityNotification(notification) && (notification.hasCommunityInvitationType() || notification.hasJoinCommunityType());
   }
 
 }
