+++ src/app/my-messages/welcome-message-card/welcome-message-card.component.html
@@ -1,3 +1,19 @@
-<p>
-  welcome-message-card works!
-</p>
+<div fxFlex>
+  <div fxLayout="row">
+    <div class="messages-description-label">Type:</div>
+    <div>{{notification.getTypeText()}}</div>
+  </div>
+  <div fxLayout="row">
+    <div class="messages-description-label">Date:</div>
+    <div>{{notification.creationDatetime | date:'dd.MM.yyyy HH:mm:ss'}}</div>
+  </div>
+  <div fxLayout="row">
+    <div class="messages-description-label">Message:</div>
+    <div>Welcome to SKOOP. Click <a href="javascript:;" (click)="openWelcomeNotificationDialog()">here</a> to get more
+      info about the system.
+    </div>
+  </div>
+</div>
+<div fxFlex="none">
+  <ng-content select="[buttonContent]"></ng-content>
+</div>
