+++ src/app/my-messages/welcome-message-card/welcome-message-card.component.ts
@@ -1,4 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { UserWelcomeNotification } from '../user-welcome-notification';
+import { InfoDialogComponent } from '../../shared/info-dialog/info-dialog.component';
+import { TemplateLoaderService } from '../../shared/template-loader.service';
+import { MatDialog } from '@angular/material';
+import { HttpErrorResponse } from '@angular/common/http';
 
 @Component({
   selector: 'app-welcome-message-card',
@@ -7,9 +12,27 @@ import { Component, OnInit } from '@angular/core';
 })
 export class WelcomeMessageCardComponent implements OnInit {
 
-  constructor() { }
+  @Input() notification: UserWelcomeNotification;
+  @Output() onErrorResponse: EventEmitter<HttpErrorResponse> = new EventEmitter();
+
+  constructor(private templateLoader: TemplateLoaderService,
+              public dialog: MatDialog) { }
 
   ngOnInit() {
+  }
+
+  openWelcomeNotificationDialog() {
+    this.templateLoader.loadTemplate('/assets/templates/welcome-notification.html')
+      .subscribe(html => {
+        this.dialog.open(InfoDialogComponent, {
+          width: '550px',
+          data: {
+            message: html
+          }
+        });
+      }, errorResponse => {
+        this.onErrorResponse.emit(errorResponse);
+      });
   }
 
 }
